2014/09/15  10:29 am.
Added some remarks concerning possible errors as well as I changed
the sign of the phase in the function calcFourierlimit.
With this sign change positive GDD leads to an delay in phase.
The final results are not effected by this change, because the
shape of the compensated function is not effected.
--> Same optimization results.